# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Street Cred API
  version: "1.0.0"
  description: |
    Server-authoritative backend for Street Cred (Trapcity).

    **Auth**: send `Authorization: Bearer <accessToken>`. Access tokens last 15 minutes; on
    `401 INVALID_TOKEN` call `/auth/refresh` with the stored refresh token and retry once.
    Refresh tokens rotate: always store the newest one.

    **Idempotency**: for actions that spend or grant currency send `Idempotency-Key: <GUID per user action>`.
    A retry with the same key returns the original response instead of charging twice.

    **Rate limits**: any endpoint can answer `429 RATE_LIMITED`; sign-in and sign-up are limited more tightly.

    **Server is authoritative**: apply the `profile` object returned by every mutating call instead of
    changing balances locally.
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://staging.example.invalid/api/v1
    description: Staging (to be provisioned)
...
security:
  - bearerAuth: []
paths:
...
  /auth/guest:
    post:
      tags: [Auth]
      security: []
      summary: Create a guest account
      description: Creates a **new** guest every call. Store the refresh token durably; it is the guest's only credential.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName: { $ref: "#/components/schemas/DisplayName" }
      responses:
        "201": { $ref: "#/components/responses/Session" }
  /auth/signup:
    post:
      tags: [Auth]
      security: []
      summary: Create an email + password account
      description: "Creates a fresh account with an empty profile. To keep a guest's progress instead, use /auth/link/email while signed in as that guest."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8, maxLength: 128 }
                displayName: { $ref: "#/components/schemas/DisplayName" }
      responses:
        "201": { $ref: "#/components/responses/Session" }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error", description: "EMAIL_TAKEN" }
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Sign in with email and password
      description: "Returns a new session. The same generic error is returned for a wrong password and an unknown email, so the endpoint cannot be used to discover registered addresses."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200": { $ref: "#/components/responses/Session" }
        "401": { $ref: "#/components/responses/Error", description: "INVALID_CREDENTIALS" }
        "403": { $ref: "#/components/responses/Error", description: "ACCOUNT_BANNED" }
  /auth/refresh:
    post:
      tags: [Auth]
      security: []
      summary: Exchange a refresh token for a new session
      description: Returns a new refresh token; the old one stops working after a 60 second grace period.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshBody" }
      responses:
        "200": { $ref: "#/components/responses/Session" }
        "401": { $ref: "#/components/responses/Error", description: "INVALID_REFRESH_TOKEN / REFRESH_TOKEN_EXPIRED / REFRESH_TOKEN_REUSED" }
  /auth/logout:
    post:
      tags: [Auth]
      security: []
      summary: Revoke one refresh token (this device)
      description: "Revokes only the refresh token you send, so other devices stay signed in. The access token remains valid until it expires (15 minutes)."
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshBody" }
      responses:
        "204": { description: Signed out }
  /auth/logout-all:
    post:
      tags: [Auth]
      summary: Revoke every refresh token for the signed-in account
      description: "Use after \"sign out everywhere\" or a suspected account compromise. Access tokens still work until they expire (15 minutes)."
      responses:
        "204": { description: Signed out everywhere }
        "401": { $ref: "#/components/responses/Error" }
  /auth/link/email:
    post:
      tags: [Auth]
      summary: Turn the signed-in guest into an email account
      description: Keeps the same user id and all progress.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                displayName: { $ref: "#/components/schemas/DisplayName" }
      responses:
        "200":
          description: Linked
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
        "409": { $ref: "#/components/responses/Error", description: "EMAIL_TAKEN / ALREADY_REGISTERED" }
  /auth/password/change:
    post:
      tags: [Auth]
      summary: Change password (signs out other devices)
      description: "Requires the current password. Every other device is signed out; this device receives a new session in the response."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string }
                newPassword: { type: string, minLength: 8 }
      responses:
        "200": { $ref: "#/components/responses/Session" }
        "401": { $ref: "#/components/responses/Error" }
  /auth/password/forgot:
    post:
      tags: [Auth]
      security: []
      summary: Request a password reset link
      description: Always returns 202 so it cannot be used to discover which emails are registered.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "202": { $ref: "#/components/responses/Ok" }
  /auth/password/reset:
    post:
      tags: [Auth]
      security: []
      summary: Set a new password with a reset token
      description: "Completes the flow started by /auth/password/forgot. The token can be used once, expires after 30 minutes, and signs out all devices."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token: { type: string }
                newPassword: { type: string, minLength: 8 }
      responses:
        "204": { description: Password changed }
        "400": { $ref: "#/components/responses/Error", description: "INVALID_RESET_TOKEN" }
  /auth/google:
    post:
      tags: [Auth]
      security: []
      summary: "Sign in with Google (not enabled yet: 501)"
      description: "Planned Google sign-in. The request shape is final, so the client can be written now; the server verifies the token and returns a session once the verifier is implemented."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [idToken]
              properties: { idToken: { type: string } }
      responses:
        "200": { $ref: "#/components/responses/Session" }
        "501": { $ref: "#/components/responses/Error", description: "NOT_IMPLEMENTED until token verification is added" }
  /auth/apple:
    post:
      tags: [Auth]
      security: []
      summary: "Sign in with Apple (not enabled yet: 501)"
      description: "Planned Apple sign-in. The request shape is final; see /auth/google for the same note."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identityToken]
              properties:
                identityToken: { type: string }
                nonce: { type: string }
      responses:
        "200": { $ref: "#/components/responses/Session" }
        "501": { $ref: "#/components/responses/Error" }
  /auth/link/google:
    post:
      tags: [Auth]
      summary: "Link Google to the signed-in account (not enabled yet: 501)"
      description: "Planned: attach a Google identity to the signed-in account so the player can also sign in that way."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [idToken]
              properties: { idToken: { type: string } }
      responses:
        "200": { description: Linked }
        "501": { $ref: "#/components/responses/Error" }
  /auth/link/apple:
    post:
      tags: [Auth]
      summary: "Link Apple to the signed-in account (not enabled yet: 501)"
      description: "Planned: attach an Apple identity to the signed-in account."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identityToken]
              properties:
                identityToken: { type: string }
                nonce: { type: string }
      responses:
        "200": { description: Linked }
        "501": { $ref: "#/components/responses/Error" }
...
  /economy/ads/reward:
    post:
      tags: [Economy]
      summary: Claim a rewarded-ad bonus (development mode only for now)
      description: "Grants the bonus after a rewarded ad. Today it works only in development mode, because a client claim is not proof that an ad was watched; production needs the ad network to confirm it server-side."
      parameters: [{ $ref: "#/components/parameters/IdempotencyKeyRequired" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rewardType]
              properties:
                rewardType: { type: string, enum: [stamina, bonus_cash] }
      responses:
        "200":
          description: Reward granted
          content:
            application/json:
              schema:
                type: object
                properties:
                  rewardType: { type: string }
                  profile: { $ref: "#/components/schemas/Profile" }
        "429": { $ref: "#/components/responses/Error", description: "AD_DAILY_CAP_REACHED" }
        "501": { $ref: "#/components/responses/Error", description: "Server-side ad verification not configured" }
  /economy/iap/verify:
    post:
      tags: [Economy]
      summary: "Verify a store receipt (placeholder: always 501)"
      description: "Where the client will send a store purchase token. Until receipt verification against Google Play and the App Store is implemented, this returns 501 and grants nothing."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [platform, productId, receipt]
              properties:
                platform: { type: string, enum: [google_play, app_store] }
                productId: { type: string }
                receipt: { type: string }
      responses:
        "501": { $ref: "#/components/responses/Error" }
  /economy/dev/founders-pass:
    post:
      tags: [Economy]
      summary: Grant the Founders Pass (development mode only)
      description: "Development shortcut for testing Founders Pass content without a real purchase. Disabled in production."
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      responses:
        "200":
          description: Granted
          content:
            application/json:
              schema:
                type: object
                properties: { profile: { $ref: "#/components/schemas/Profile" } }
        "409": { $ref: "#/components/responses/Error", description: "ALREADY_OWNED" }
        "501": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Recommended. A GUID per user action; retries with the same key return the first response.
      schema: { type: string, minLength: 8, maxLength: 128 }
    IdempotencyKeyRequired:
      name: Idempotency-Key
      in: header
      required: true
      description: Required. A GUID per user action; retries with the same key return the first response.
      schema: { type: string, minLength: 8, maxLength: 128 }
  responses:
    Ok:
      description: Success
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
    Error:
      description: Failure
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Session:
      description: Signed in
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Session" }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: INSUFFICIENT_FUNDS }
            message: { type: string }
            details: { description: Optional context, e.g. failing fields or availableAt }
    DisplayName:
      type: string
      minLength: 3
      maxLength: 24
      description: Letters, numbers, spaces and `_ . -`
    RefreshBody:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }
...
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: [string, "null"] }
        displayName: { type: string }
        isGuest: { type: boolean }
        role: { type: string, enum: [player, admin] }
        createdAt: { type: string, format: date-time }
    Stamina:
      type: object
      properties:
        current: { type: integer }
        max: { type: integer }
        regenSeconds: { type: integer }
        nextPointAt: { type: [string, "null"], format: date-time }
        fullAt: { type: [string, "null"], format: date-time }
    Profile:
      type: object
      description: The authoritative player state. Replace local values with this on every response.
      properties:
        level: { type: integer }
        xp: { type: integer, description: Progress inside the current level }
        xpToNextLevel: { type: integer }
        totalXp: { type: integer }
        cash: { type: integer }
        gems: { type: integer }
        stamina: { $ref: "#/components/schemas/Stamina" }
        equippedVehicleId: { type: [string, "null"] }
        equippedOutfitId: { type: [string, "null"] }
        hasFoundersPass: { type: boolean }
    Session:
      type: object
      properties:
        accessToken: { type: string }
        accessTokenExpiresIn: { type: integer, example: 900 }
        refreshToken: { type: string }
        refreshTokenExpiresAt: { type: string, format: date-time }
        user: { $ref: "#/components/schemas/User" }
        profile: { $ref: "#/components/schemas/Profile" }
...